+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@ Release Notes
 
 ## 1.7.27
 * Action Groups: Adds support for action groups to perform actions when alerts fire.
+* Functions: AzureWebJobsDashboard setting is only added for a V1 Function app
 * Functions: Add the correct netFrameworkVersion to the template based on the runtime version
 
 ## 1.7.26